requirements.txt
pytest==6.2.2
selenium==3.141.0
pytest-bdd==4.0.2
pytest-html==3.1.1
validators==0.18.2
Pillow==8.3.0